move cpp_audio out of src and update build
