+++ README_Audio.md
@@ -74,7 +74,7 @@ This behaviour ensures that very long steps do not delay the preview yet short
 steps can still be heard in full.
 
 ## C++ Port
-A minimal C++ implementation using JUCE lives in `src/cpp_audio`. Build it with CMake and ensure JUCE is available on your system.
+A minimal C++ implementation using JUCE lives in `cpp_audio`. Build it with CMake and ensure JUCE is available on your system.
 
 The JUCE port now provides a `StepPreviewComponent` that mirrors the Python UI's play/pause/stop controls and time slider for auditioning a single step. It also includes a **Reset** button and labels the currently loaded step alongside the playback time.
