+++ cpp_audio/BufferAudioSource.h
