+++ cpp_audio/Common.cpp
