+++ cpp_audio/SynthFunctions.h
