+++ cpp_audio/presets/NoiseParams.cpp
