+++ cpp_audio/synths/HybridQamMonauralBeat.cpp
