+++ cpp_audio/synths/MonauralBeatStereoAmps.cpp
