+++ cpp_audio/synths/QamBeat.cpp
