+++ cpp_audio/synths/RhythmicWaveshaping.cpp
