+++ cpp_audio/synths/SpatialAngleModulation.cpp
