+++ cpp_audio/synths/StereoAMIndependent.cpp
